+++ sdk/eventhubs/azure-messaging-eventhubs/src/event_data_batch.cpp
@@ -14,7 +14,7 @@ using namespace Azure::Core::Diagnostics;
 
 namespace Azure { namespace Messaging { namespace EventHubs {
 
-  bool EventDataBatch::TryAddMessage(Azure::Messaging::EventHubs::Models::EventData const& message)
+  bool EventDataBatch::TryAdd(Azure::Messaging::EventHubs::Models::EventData const& message)
   {
     return TryAddAmqpMessage(message.GetRawAmqpMessage());
   }
@@ -77,8 +77,17 @@ namespace Azure { namespace Messaging { namespace EventHubs {
     auto actualPayloadSize = CalculateActualSizeForPayload(serializedMessage);
     if (m_currentSize + actualPayloadSize > m_maxBytes.Value())
     {
-      m_currentSize = 0;
-      m_batchEnvelope = nullptr;
+      Log::Stream(Logger::Level::Informational)
+          << "Batch is full. Cannot add more messages. "
+          << "Message size: " << actualPayloadSize << " size: " << m_currentSize
+          << " Max size: " << m_maxBytes.Value() << std::endl;
+      // If we don't have any messages and we can't add this one, then we can't add it at all.
+      // Discard the contents of the batch.
+      if (m_marshalledMessages.size() == 0)
+      {
+        m_currentSize = 0;
+        m_batchEnvelope = nullptr;
+      }
       return false;
     }